Remove BindableView.JavaFinalize
The dispatched call to remove views from the overridden Java finalize() causes sporadic crashes due to modifying already-finalized views. Remove for now until a better approach can be found.

This is a non-breaking change because the current implementation doesn't do anything, due to a bug where removing views in this way doesn't change the DependencyObjectStore.Parent of the removed view (hence view never gets released to the template pool as intended).