Implement Page pooling
Allow Pages to be pooled in the same way that FrameworkTemplate instances are currently pooled. This can be a significant performance improvement in certain Frame navigation scenarios.

For now Pages are released to the pool conservatively (only on New navigation). It should be possible to optimize more aggressively by releasing pages when consumers modify the Back- and ForwardStacks programmatically.