Phase binding - prevent access of disposed bindings
Ensure that BindingPath updates managed weakRef of child BindingItem if it disposes the old weakRef, even if the target hasn't changed.

This prevents a crash caused by the BindingItem retaining the disposed weakRef in certain situations where Phase binding is used on a chained binding and a list's ItemsSource is modified multiple times.