[Android] Prevent sibling of transformed view stealing focus
The new dispatchTouchEvent() code to handle children with StaticTransformations calls child.dispatchTouchEvent() for all children, unlike ViewGroup.dispatchTouchEvent() which only calls for children for whom the touch lies in bounds. At least one consequence of this is requestFocus() wrongly getting called in some circumstances.

As a narrow workaround, we block requestFocus() from being propagated in this specific case.