[iOS] Apply styles root-first on adding view to live tree
This fixes a bug where Loading/Loaded was called twice for some views, in the case that they're inside the Content of a control whose template changes when the style is applied.

This is necessary for iOS specifically because Loading/Loaded are raised 'leaf first' by the native platform, unlike the other Uno targets.