[iOS] Fix bug where list requests wrong height
When deciding whether to erase the layout, don't do it if the new available size matches that used for the most recent arrange.

This fixes bug occurring after a complicated sequence of measure and arrange calls where the list would drop the databound item measurements, but PreferredLayoutAttributesFittingAttributes() would not be re-called.