[iOS] Call UIElement.InvalidateMeasure() when Visibility changes
Calling this method instead of SetNeedsLayout means that OnInvalidateMeasure() is called for types that override it.

This fixes a bug introduced by #1726 (somehow) whereby changing TextBlock.Foreground and setting the TextBlock to Collapsed in the same synchronous pass would not result in the new Foreground showing when the TextBlock was made visible again.