[iOS] Fix bug with ListViewItem.Margin application
When a view requires measure in LayoutSubviews(), add its Margin to its Bounds to determine the available size. This fixes a bug where the contents of a ListViewItem could be squished by the size of its Margin.

This is required because of recent changes where Margin is handled by views themselves, rather than their parent layouter.