[Android] Don't apply StretchAffectsMeasure with infinity
This was causing content to measure itself as 0 if measured dimension was infinite and StretchAffectsMeasure is true (ie if parent is non-DependencyObject). This was causing some of the samples in CommandBar_LongTitle not to display, for which an incorrect fix was applied.

(cherry picked from commit a4ccfc75f819720227f7e723626ecc241ebdbc81)