ListView - fix stack overflow when changing SelectedItem
When changing SelectedItem to an invalid value, if the current value had already become invalid, an infinite ping-pong from one value to the other would ensue.

This change interrupts the process by setting the value to null instead, which is a little different from Windows where the original value is retained even if invalid.