[Android] Include Margin when measuring from LayoutManager
The LayoutManager is used when AreDimensionsConstrained = true for a particular view, which can be set manually as an optimization if it's known that a view's size will never change. It wasn't taking the Margin into account when measuring the view, which was causing errors after layouting changes.

(cherry picked from commit 1eac9eb8dbf278b5b7918f8415772bdad144489b)