fix: VisibleBoundsPadding causing text alignment problem on iOS.
Looks like a concurrency issue in iOS itself with `NSTextContainer`: at a very specific measurement, when the text is multi-line (max-lines=2), center-aligned and there's a notch (iPhone-X), it's causing a very weird alignment problem of the text where it's centered on the right-hand bound of the container, cutting the text in half.

The solution consists to requeue the `.SetPadding()` in the `VisibleBoundsPadding` (for iOS only). This could cause flickering issue which need to be tested, but should solve the reported problem.