fix(android): Fix for bug https://github.com/unoplatform/uno/issues/2879
The problem was caused by a cache in Android layout system. Was appearing when the `DesiredSize` is required to layout an element, like in the `<StackPanel>`. In this case, the `Measure()` were not properly called at the right time, causing the panel (`<StackPanel>` here) to use the previous (wrong) measured/desired size.

The fix is to call `ForceLayout()`, causing Android to bypass the cache during the `Measure()` phase.

(cherry picked from commit c4dd71391f160a33689e46c40dce14d4070c42b8)