fix(elevatedview): Make sur the `ElevatedView` recopy its `TemplatedParent`.
When the `ElevatedView` itself was placed into an `ControlTemplate`, the `TemplatedParent` were not the right one, leading to unexpected behavior in the result.

This is a patch and reproduce a behavior already patched in `ContentControl`. A better fix would required architectural changes in Uno, which is not is the scope of this bug fix.