test(runtime): Try to improve stability of RuntimeTests
fixes #2894

The RuntimeTests UI test frequently polls the running test state using GetDependencyPropertyValue(). For some reason, on some platforms (Android and iOS) the inner invocation occasionally returns no results in this particular context.

We don't really care, so to make it more robust, use the non-generic version of GetDependencyPropertyValue() which calls FirstOrDefault() instead of First() and thus doesn't throw when this happens.