feat(scrollcontentpresenter)!: [Android] [iOS] [MacOS] Make ScrollContentPresenter inherit from ContentPresenter
BREAKING CHANGE: ScrollContentPresenter on Android/iOS/MacOS is now a FrameworkElement, instead of inheriting directly from native scroll viewer. The native scroll viewer is now a child of the managed ScrollContentPresenter.

On Android only, the NativeScrollContentPresenter continues to implement IFrameworkElement, since it uses the Layouter to arrange its children which is tightly coupled to IFrameworkElement.