chore: [WASM] ListView - handle insertion when scrolled to end
Don't try to scroll in response to observable collection changes when there's no space to scroll. This still results in a jump in the scroll position, but is better than being left in an unrecoverable bad layout state. The underlying challenge is that the panel size as seen by the ScrollViewer only changes after a measure/arrange pass, which occurs after ChangeView() is called.