fix(revealbrush): [Android] Fix exception when using RevealBrush
Implement Brush.GetPaintInner() to use FallbackColor or Color when RevealBrush set. This is not a complete implementation even of fallback flat color (eg it doesn't subscribe to changes of Color, and some places in code need to be explicitly modified to accept new types of brushes), it's just to avoid an exception if a RevealBrush is used.