fix: Some controls might be recycled on creation
This commit removes a fix (https://github.com/unoplatform/uno/commit/fe83e99a  - https://github.com/unoplatform/uno/pull/607) that was made to avoid adding the same element twice in the visual tree.
We tried to run the sample that was made to validte this fix, and apprently it is no longer relevant.