chore: Only use materialized template root when it's a container
The previous approach was too intrusive and had observable side-effects (SelectorItem.Content was not the right object). We no longer try to inject the template root if it's not a container, but we cache the knowledge that it's not and return it to the pool, so worst-case it's cheap and best-case it's free.