fix: Fix Slider thumb not visible when Value is at maximum
There was a bug occurring on iOS where the indicator Rectangle was having its size set from within SizeChanged before the Thumb had a chance to be arranged, leaving it to take the entire space and leave none for the Thumb.

As a workaround, return the Width/Height if ActualWidth/ActualHeight haven't been set yet.