fix: Don't propagate DataContext via Border.Child
Restore the Uno 2.x behaviour whereby Border.ChildProperty doesn't propagate inherited values (including DataContext) to the property value. The inheritance is handled by the visual tree in the normal case (since Child is always a view).

This restores the ability for the framework to call SetParent() to simulate a different inheritance hierarchy than the actual visual tree configuration, eg for AppBarButton using the native style.