fix(contentdialog): [Android] Fix botom of full-size ContentDialog cut off
Add an internal adjusted visible bounds which trims the status bar from the height. This allows the ContentDialog to be arranged with the correct available height.

There's a case for adjusting the public-facing VisibleBounds property, but it's not clear what the exact adjustment would be and it's potentially a breaking change, and isn't required for the present case.