fix(listview): [iOS] Fix outer ElementName bindings not available on reload
Bindings by ElementName to an element outside of a ListViewItem template were getting detached in some cases after the list was unloaded from the window and reloaded, because the ListViewItem's logical Parent was incorrectly getting set to the visual-tree parent (which is a native UIView imposed by UICollectionView). Ensure the logical Parent is correctly restored.