fix(xaml): Resolve markup extensions in Xaml correctly
Take the namespace into account when testing for a MarkupExtension declaration in Xaml. This brings behavior closer to UWP, and fixes a regression in Uno 3.2 where usage of a different type could be mistaken as the shortened form of a markup extension (eg FontIcon for FontIconExtension).

(cherry picked from commit 667fdee9416e5c4ebdfd4bc38ca0c8a6a3bcfdba)