fix(xaml): Ensure reference equality of x:Name resource by different access methods
Don't assign to generated field twice (just ensure that any lazy initializer in the ResourceDictionary runs). This ensures that the resource accessed via the field is reference-equal to that set by StaticResource, which is required for the common scenario of defining a CollectionViewSource in Xaml and setting its Source in code-behind.