fix(Button): Touch pressed do not update Visual State
Using a touch pointer, touch screen, with Skia GTK does not set the
IsPointerOver only the IsPointerPressed when touching a button. The
check for IsPointerPressed and IsPointerOver is never true for the
touch case, so the Visual State does not change the appearance when
pressed.

I believe that the normal thing is that if we have an IsPointerPressed
true automatically IsPointerOver is also true, I cannot think of any
case where this is not true, but for touch this is getting only the
pressed not the over, please correct me if I am wrong. This commit
removes the IsPointerOver from the check to set the visual state to
"Pressed".

Signed-off-by: Matheus Castello <matheus@castello.eng.br>