fix(GTK): pointer events in wrong locations using Wayland
For GTK applications using Wayland the title bar is rendered by the
window itself and not by the composer. And it seems to me that GTK
counts the title bar as part of the window and uses it to trigger the
window events, leading to erroneous considerations about the size and
location of pointers in the area where it is used to draw the
components.

The fix adds a GtkEventBox in the window, instead of use the window
events we will be based on the events of the GtkEventBox which we can
assure that the GTK will take into account the title bar.

Related-to: #5706
Signed-off-by: Matheus Castello <matheus@castello.eng.br>