fix(menuflyout): Position MenuFlyout correctly with native AppBarButton
When native AppBarButton style is being used in conjunction with XAML MenuFlyout style, ensure that the native app bar button view is used as anchor when positioning the flyout, because the managed AppBarButton is not actually in the visual tree.

Note: if MenuFlyout.UseNativePopup is set to "True" (native flyout style), the positioning is already correctly handled.