perf: Apply various workarounds for dotnet/runtime issue #56309
These changes are required to reduce the interpreter performance hit.  The AOT compiler detects try/catch/finally containing methods and forces the interpreter, even if the try/catch is not in the execution path.

Inner methods are created to reduce the size of the interpreter-executed code, and return to AOT as soon as possible.