feat(listview): [iOS] [Android] Support ItemsPresenter.MinHeight
Match Windows' behavior when ItemsPresenter.MinHeight (or MinWidth for horizontal-scrolling list) is set inside of a ListView or GridView. Since the ItemsPresenter is actually outside the scrollable container (NativeListViewBase), this is implemented by mapping the MinHeight value to the native layout calculated by VirtualizingPanelLayout.

This supports scenarios where a minimal scrollable extent, regardless of number of items, is desired.