Update get-started-rider.md
I've tried quite a few ways to try to use Rider for Android development from a linux box and there don't seem to be readily available instructions that work and references to this setup working seem to all be labelled obsolete as of around 2018. I can't seem to get Xamarin.Android to build on Ubuntu 20.04 at all (I think it might be an incompatibility between mono and the .net 6 sdk maybe...? the Xamarin.Android build system is also failing on linux at the moment.) This very well could be just me not figuring something out but people should at least be aware of the issue before they start pursuing it. Uno-check also seems to skip android on linux which makes troubleshooting this incompatibility very frustrating. This is my proposed stab at an update to the documentation on the issue.  If this change is accepted it should be done in all locations where the support grid appears though I figured I'd just do one proposed change so the wording can be tweaked before I put in any other proposed changes.