fix(flipview): [Android] Fix IllegalStateException from FlipView
Fix FlipView.SelectedIndex being reset to 0 whenever OnItemsChanged() is raised. This was a hangover from a much earlier state of the code where the method was only raised the first time the Items property was set; aside from being incorrect,  it was causing exceptions on Android due to trying to scroll the native pager control before updating its state.