chore(hotreload): Add flags to force enable or disable hot reload
Add a project flag to force settings for XAML-generated hot reload callbacks (which are normally governed by whether the build is a debug build), and a feature configuration flag to force-disable hot reload at runtime. This allows hot reload-related scenarios to be tested alongside the normal no-hot-reload scenario in unit tests, and the feature flag additionally allows app developers to disable hot reload for any reason.