fix(listview): Fix blank ListView inside ListView
Fix bug where ListView with non-virtualizing items panel inside another ListView could be blank, because the list materializes the item template to check if the root is a ListViewItem, and returning the template root to the pool was causing the native scroll content presenter to remove its content. We fix this by disabling element clean-up from the pool in this case, since it's not necessary.