fix: ContentDialog does not show up on startup on WPF
The problem here is caused by the fact that the ContentDialog layout is initialized only when the dialog is first created, which is a problem if the window is very small (or 0x0 in case of startup on WPF). The changes make it respond to Window.SizeChanged which fixes these scenarios.