fix: ContentPresenter content alignment
Previously content alignement properties on ContentPresenter were overriding the alignment of the child, which did not work properly in many cases (especially when the alignment was directly set on the content element). Instead ContentPresenter should arrange the content based on the content alignment properties and not affect the actual alignment of the content element.