fix(TextBox): [iOS] Add temporary workaround for iOS 16 on selection with UITextField
This change is needed to make the SinglelineTextBoxView.SelectedTextRange conditional, and avoid changing the value provided to UIKit. This breaks the use of `TextBoxSelectionChanged`, but can be restored by using the `FeatureConfiguration.TextBox.IOS16EnableSelectionSupport` feature flag if the issue ends up being fixed in UIKit. See https://github.com/unoplatform/uno/issues/9430 for more details.