fix(reg): Remove invalid removal of uno.ui files when building with WinAppSDK.
The current change removes all binaries from the net5.0-windows target, which is incorrect as it also removes uno.toolkit.ui (as it includes ElevatedView and VisibleBoundsPadding).

The proper fix is to specifically modify Uno.WindowsCommunityToolkit, which does not include a Windows target. This causes netstandard2.0 binaries to incorrectly show up in the build.

Related to https://github.com/unoplatform/uno/issues/10182