fix: System.IO.DirectoryNotFoundException looking for /dev/input
When using it on a device with no input, very common in headless
applications, an exception is thrown preventing the program from
running. A workround would ask to always have a mouse or keyboard
connected, which is not feasible in some use cases. So, let's just
check if the directory exists before trying to open it and life
goes on.

Signed-off-by: Matheus Castello <matheus.castello@toradex.com>