fix: bump minimum macOS version for SampleApp
```
MacOS/Info.plist : error : The LSMinimumSystemVersion value in the Info.plist (10.11) does not match the SupportedOSPlatformVersion value (10.14) in the project file (if there is no SupportedOSPlatformVersion value in the project file, then a default value has been assumed). Either change the value in the Info.plist to match the SupportedOSPlatformVersion value, or remove the value in the Info.plist (and add a SupportedOSPlatformVersion value to the project file if it doesn't already exist). [/xxx/uno/src/SamplesApp/SamplesApp.netcoremobile/SamplesApp.netcoremobile.csproj::TargetFramework=net7.0-macos]
```