fix: Adjust support for single-instance generator
The change take into account the fact that generators may be instanciated only once per solution build, causing incorrect reuse of generation run across the solution. Runs and hashes are now segmented per targetframework and project file.

The clearing of runs is also restored as a new build initiated by VS needs to be detected as the base line for hot reload.