fix: Fix ListView missing top elements when flinging #11809
From logs, when isLargeScroll==true and ScrollOffset is 0, _dynamicSeedIndex would become
1 (index == 0, sign = -1, 0 - (-1) = 1). The index path is then passed
to FillForward's GetNextUnmaterializedItem. The next item thus would
become the 2nd or 3rd items, which is incorrect. The correct item should be the 1st.

The fix is to add a condition on index == 0 and pretend
currentMaterializedItem is -1. The next item would then correctly be 0.