fix: Ensure theme is correctly applied to elements constructed before OnLaunched
It is possible to construct some UI in the app constructor. Because of that, we need to make sure the default OS-based app theme is initialized immediately, before the constructor is even running. In addition, we set the "last used theme" when default Foreground is set to make sure that it is correctly updated when first attached to the window.