fix: Avoid propagating routed events up from PopupPanel
In case of UWP and WinUI, routed events are not propagated up from Popups, even if they are part of the visual tree and their Parent is set. Because we currently override the logical parent on Android and iOS by overwriting the actual parent, the only way to stop the propagation is to actually stop it at the PopupPanel level.

(cherry picked from commit 326a0eca234510856cf7aa83e7c6fcd093daad7a)