fix(hr): Ensure that the msbuild version used is properly aligned
Ensures that the MSBuild version is determined using the appropriate folder, where `dotnet --version` used the latest available SDK on the machine. With this change, the MSBuild version will be the one explicitly used by the app. Combined with https://github.com/unoplatform/uno/pull/14170, the devserver now completely uses the appropriate SDK/runtime version based on the app's compilation.