fix(skia): Fix ineffective FillRule setting for StreamGeometry on Skia platform
This PR fixes the issue of ineffective FillRule setting for StreamGeometry on the Skia platform. The root cause is that the FillRule of StreamGeometry was not assigned to SKPath on the Skia platform, which resulted in SKPath always maintaining the default value. This would cause the Path's Data corresponding to the XAML code like `<Path Data="xxxxx"/>` used on the Skia platform to use the Nonzero FillRule method, instead of maintaining the default EvenOdd FillRule method consistent with WinUI.