fix(gtk): Fix issue where setting background color in OnLaunched is ineffective on Skia.Gtk platform
This PR fixes an issue where setting the background color immediately in OnLaunched using the MainWindow.SetBackground method is ineffective on the Skia.Gtk platform.

The root cause of the issue is that when the MainWindow.SetBackground method is used immediately in OnLaunched, it enters UnoGtkWindowHost.UpdateRendererBackground. At this point, the `_renderer` field has not been initialized, so no action is taken. It is only after InitializeAsync executes GtkRendererProvider.CreateForHostAsync that the `_renderer` field is obtained. However, by this time, the logic for assigning the background color has been missed.

The fix is to assign the `_renderer.BackgroundColor` after the execution of GtkRendererProvider.CreateForHostAsync method.

Fix https://github.com/unoplatform/uno/issues/15988