chore: now begins the hacking…
I'm still trying to figure out what's up with the error:

    Unhandled exception: Command workload has more than one child named "--info".

I can't repro that error locally, nor can I make
`System.CommandLine` throw it deliberately.

So as a Glorious Hack™, what if we just *replaced*
`System.CommandLine.dll` with a copy from .NET 10 beta 5?
(Not that I have any reason to believe that will be better, and the
version in .NET 10 Preview 5 *also* has the change which *could*
throw that exception…)

I'm running out of ideas.