chore: fix local-android-uitest-run.sh et al to work on macOS
Export and use `$ANDROID_AVD_HOME`, so that we have a "known" location
for whare Emulator images are created.  (Not sure what's wrong with
my environment, but they are not reliably created or loaded from
`$HOME/.android/avd`.)

Update `local-android-uitest-run.sh` so that it doesn't require running
the script from the `build/test-scripts` directory.  You can now run it
from topdir:

	UITEST_IS_LOCAL=true \
	  build/test-scripts/local-android-uitest-run.sh

Update the expected paths to match .NET 9 output.