chore: workaround extra array wrapping
The good news is that the `SamplesApp.Skia.netcoremobile` unit tests
are now executing within an Android+NativeAOT environment.

The bad news is that half of them are failing:

	System.Reflection.TargetParameterCountException: Arg_ParmCnt
	    at System.Reflection.DynamicInvokeInfo.ThrowForArgCountMismatch() + 0x83
	    at System.Reflection.DynamicInvokeInfo.Invoke(Object, IntPtr, Object[], BinderBundle, Boolean) + 0x136
	    at Internal.Reflection.Execution.MethodInvokers.InstanceMethodInvoker.Invoke(Object, Object[], BinderBundle, Boolean) + 0x4f
	    at Internal.Reflection.Core.Execution.MethodBaseInvoker.Invoke(Object, Object[], Binder, BindingFlags, CultureInfo) + 0x48
	    at System.Reflection.Runtime.MethodInfos.RuntimeMethodInfo.Invoke(Object, BindingFlags, Binder, Object[], CultureInfo) + 0x70
	    at Uno.UI.Samples.Tests.UnitTestsControl.<>c__DisplayClass67_2.<<ExecuteTestsForInstance>g__InvokeTestMethod|5>d.MoveNext() + 0x4b9

	System.InvalidOperationException: Missing parameter does not have default value
	    at Uno.UI.Samples.Tests.UnitTestsControl.ExpandArgumentsWithDefaultValues(Object[] methodArguments, ParameterInfo[] methodParameters) + 0x138
	    at Uno.UI.Samples.Tests.UnitTestsControl.<>c__DisplayClass67_4.<<ExecuteTestsForInstance>b__12>d.MoveNext() + 0x92
	--- End of stack trace from previous location ---
	    at Uno.UI.Samples.Tests.UnitTestsControl.<>c__DisplayClass67_2.<<ExecuteTestsForInstance>g__InvokeTestMethod|5>d.MoveNext() + 0x549

Initial investigation suggests that [`ITestDataSource.GetData()`][0]
is wrapping data in an extra array, i.e. if in a non-Android+NativeAOT
environment it provides `{1, 2}`, in Android+NativeAOT it instead
provides `{ {1, 2} }`.

Update the `UnitTestMethodInfo()` constructor to check for this extra
wrapping, log if it is detected, and return the inner data.  This
*should* allow more unit tests to execute.

Additionally, this extra wrapping only happens if the test method
takes more than one parameter.  Don't check for extra wrapping for
test methods accepting only one parameter.

TODO: figure out *why* this is happening on Android+NativeAOT, file
bugs where appropriate.

[0]: https://learn.microsoft.com/en-us/dotnet/api/microsoft.visualstudio.testtools.unittesting.itestdatasource.getdata?view=mstest-net-3.9