chore: ignore IL2026 in BindableMetadata.g.cs
Context: 559b1f7866028a1d23f74ba43c3e8614a8e9f284

Commit 559b1f78 updated `BindableMetadata.g.cs` to ignore IL2111,
because `typeof(Type)` in the binding code was emitting it.

Turns out, `typeof(Excpetion)` emits a similar IL2026!

	Error: …/RecommendedMarkup/RecommendedMarkup/obj/Debug/net10.0-browserwasm/Uno.UI.SourceGenerators/Uno.UI.SourceGenerators.BindableTypeProviders.BindableTypeProvidersSourceGenerator/BindableMetadata.g.cs(1086,4): error IL2026:
	  Using member 'System.Exception.TargetSite.get' which has 'RequiresUnreferencedCodeAttribute' can break functionality when trimming application code. Metadata for the method might be incomplete or removed.

which is flagged on this line of code:

	bindableType.AddProperty("Error", typeof(global::System.Exception), GetError);

This was encountered from uno.templates CI when building the output of:

	dotnet new unoapp -n RecommendedMarkupDsp -skip -preset recommended -markup csharp -dsp

Update `BindableTypeProvidersGenerationTask` so that IL2026 is ignored.